requirements.txt
# --- AI + Monitoring ---
openai>=1.40.0
langfuse>=2.0.0
langchain>=0.2.0

# --- Streamlit ---
streamlit==1.33.0
Pillow>=9.0.0

# --- ML / Model ---
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.2
joblib>=1.3.0

# --- DigitalOcean Spaces / S3 ---
boto3==1.34.131
botocore==1.34.131
s3fs==2024.6.1

# --- Utilities ---
python-dotenv==1.0.1
